requirements.txt
mistralai>=1.0.0
python-substack-tarun>=0.2.3
feedparser>=6.0.11
requests>=2.32.0
python-frontmatter>=1.1.0
jinja2>=3.1.4
